Setup users blueprint
