+++ services/users/manage.py
@@ -1,9 +1,11 @@
 import unittest
 from flask.cli import FlaskGroup
 
-from project import app, db
+from project import create_app, db
+from project.api.models import User
 
-cli = FlaskGroup(app)
+app = create_app()
+cli = FlaskGroup(create_app=create_app)
 
 
 @cli.command()
@@ -11,7 +13,6 @@ def recreate_db():
     db.drop_all()
     db.create_all()
     db.session.commit()
-    print("Done")
 
 
 @cli.command()
