+++ services/users/project/__init__.py
@@ -1,31 +1,26 @@
 import os
-import sys
-from flask import Flask, jsonify
+
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
-app = Flask(__name__)
 
-app_settings = os.getenv('APP_SETTINGS')
-app.config.from_object('project.config.DevelopmentConfig')
-
-db = SQLAlchemy(app)
+db = SQLAlchemy()
 
 
-class User(db.Model):
-    __tablename__ = 'users'
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    username = db.Column(db.String(128), nullable=False)
-    email = db.Column(db.String(128), nullable=False)
-    active = db.Column(db.Boolean(), default=True, nullable=False)
+def create_app(script_info=None):
 
-    def __init__(self, username, email):
-        self.username = username
-        self.email = email
+    app = Flask(__name__)
 
+    app_settings = os.getenv('APP_SETTINGS')
+    app.config.from_object(app_settings)
 
-@app.route('/users/ping', methods=['GET'])
-def ping_pong():
-    return jsonify({
-        'status': 'success',
-        'message': 'pong'
-    })
+    db.init_app(app)
+
+    from project.api.users import users_blueprint
+    app.register_blueprint(users_blueprint)
+
+    @app.shell_context_processor
+    def ctx():
+        return {'app': app, 'db': db}
+
+    return app
