+++ services/users/project/api/users.py
@@ -1,0 +1,12 @@
+from flask import Blueprint, jsonify
+
+
+users_blueprint = Blueprint('users', __name__)
+
+
+@users_blueprint.route('/users/ping', methods=['GET'])
+def ping_pong():
+    return jsonify({
+        'status': 'success',
+        'message': 'pong'
+    })
