+++ services/users/project/tests/base.py
@@ -1,10 +1,11 @@
 from flask_testing import TestCase
-from project import app, db
+from project import create_app, db
 
 
 class BaseTestCase(TestCase):
 
     def create_app(TestCase):
+        app = create_app()
         app.config.from_object('project.config.TestingConfig')
         return app
